+++ apps/portal/src/components/widgets/staking/Providers/hooks/nominationPools/useNominationPoolsProviders.ts
@@ -1,30 +1,17 @@
 import useAprs from './useAprs'
+import useAvailableBalances from './useAvailableBalances'
 import useUnlockDurations from './useUnlockDurations'
-import { selectedSubstrateAccountsState } from '@/domains/accounts'
 import { nominationPoolsEnabledChainsState } from '@/domains/chains'
-import { chainDeriveState } from '@/domains/common'
 import { Decimal } from '@talismn/math'
-import { useMemo } from 'react'
-import { useRecoilValue, waitForAll } from 'recoil'
+import { useRecoilValue } from 'recoil'
 
 const useNominationPoolsProviders = () => {
   const nominationPools = useRecoilValue(nominationPoolsEnabledChainsState)
-  const accounts = useRecoilValue(selectedSubstrateAccountsState)
-  const addresses = useMemo(() => accounts.map(x => x.address), [accounts])
   const rpcIds = nominationPools.map(({ rpc }) => rpc ?? '')
-
-  const balancesByRpc = useRecoilValue(
-    waitForAll(rpcIds.flatMap(apiId => addresses.map(address => chainDeriveState(apiId, 'balances', 'all', [address]))))
-  )
 
   const aprs = useAprs({ rpcIds })
   const unlockDurations = useUnlockDurations({ rpcIds })
-
-  const reducedBalancesByRpc = rpcIds.map((_, index) => {
-    // groups balances of multiple addresses by rpc
-    const balancesForRpc = balancesByRpc.slice(index * addresses.length, (index + 1) * addresses.length)
-    return balancesForRpc.reduce((prev, curr) => prev + curr.availableBalance.toBigInt(), 0n)
-  })
+  const availableBalances = useAvailableBalances({ rpcIds })
 
   const nominationPoolProviders = nominationPools?.map(({ chainName, id, nativeToken }, index) => {
     const { symbol, logo, decimals } = nativeToken ?? {}
@@ -37,7 +24,7 @@ const useNominationPoolsProviders = () => {
       type: 'Nomination pool',
       provider: chainName,
       unbondingPeriod: unlockDurations[index],
-      availableBalance: Decimal.fromPlanck(reducedBalancesByRpc[index] ?? 0n, decimals ?? 0, {
+      availableBalance: Decimal.fromPlanck(availableBalances[index] ?? 0n, decimals ?? 0, {
         currency: symbol,
       }).toLocaleString(),
       stakePercentage: 0.5,
